requirements.txt
fastapi==0.135.1
python-dotenv==1.2.2
requests==2.32.5
uvicorn==0.42.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.20

# Testing
httpx==0.28.1
pytest==8.3.5
